fix species update and add test for species
